Removing the Collection model and moving all functions into Group
Also renaming collection_id column in group options to facilitate the relationship moves

This does not yet rename the table
