+++ app/models/group_option.rb
@@ -4,7 +4,7 @@ class GroupOption < ApplicationRecord
 
   EMAIL_MESSAGES = 0
 
-  belongs_to :group, class_name: "Group", foreign_key: "collection_id"
+  belongs_to :group, class_name: "Group"
   belongs_to :user
 
   enum option_type: { email_messages: 0 }
