+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2023_05_16_133007) do
+ActiveRecord::Schema.define(version: 2023_05_16_194040) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -46,9 +46,9 @@ ActiveRecord::Schema.define(version: 2023_05_16_133007) do
   create_table "collection_options", force: :cascade do |t|
     t.integer "option_type"
     t.integer "option_value"
-    t.bigint "collection_id"
+    t.bigint "group_id"
     t.bigint "user_id"
-    t.index ["collection_id"], name: "index_collection_options_on_collection_id"
+    t.index ["group_id"], name: "index_collection_options_on_group_id"
     t.index ["user_id"], name: "index_collection_options_on_user_id"
   end
