--- spec/factories/collection.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-FactoryBot.define do
-  factory :collection do
-    title { FFaker::Skill.tech_skill }
-    code { FFaker::Code.npi }
-  end
-end
